move more methods to separate classes
