+++ png-prototype/read.php
@@ -8,6 +8,8 @@ use Mike42\GfxPhp\RgbRasterImage;
 use Mike42\GfxPhp\IndexedRasterImage;
 use Mike42\GfxPhp\Codec\Common\DataInputStream;
 use Mike42\GfxPhp\Codec\Common\DataBlobInputStream;
+use Mike42\GfxPhp\Codec\Png\PngHeader;
+use Mike42\GfxPhp\Codec\Png\PngChunk;
 
 /**
  * Takes 8-bit samples, and produces eight times as many 1-bit samples,
@@ -182,215 +184,6 @@ function unfilter(array $currentFiltered, array $prior, int $filterType, int $bp
         return $ret;
     }
     throw new Exception("Filter type $filterType not valid");
-}
-
-/**
- * Reading PNG files in PHP.
- */
-class PngChunk
-{
-    private $type;
-    private $data;
-    private $crc;
-
-    public function __construct(string $type, string $data)
-    {
-        $this -> type = $type;
-        $this -> data = $data;
-        // Always compute CRC based on the data we have.
-        // If this is being read from a chunk's binary, then
-        // this will be compared, if not, it will be written.
-        $this -> crc = crc32($type . $data);
-    }
-
-    public function toBin()
-    {
-        $len = strlen($this -> data);
-        $lenData = pack("N", $len);
-        $bodyData = $this -> type . $this -> data;
-        $crcData = pack("N", $this -> crc);
-        return $lenData . $bodyData . $crcData;
-    }
-  
-    public function getCrc()
-    {
-        return $this -> crc;
-    }
-
-    public function getType()
-    {
-        return $this -> type;
-    }
-  
-    public function getData()
-    {
-        return $this -> data;
-    }
-
-    public static function isValidChunkName(string $name)
-    {
-        if (array_search($name, ["IHDR", "IDAT", "PLTE", "IEND"], true) !== false) {
-            // Critical chunks defined as of PNG 1.2
-            return true;
-        } else if (preg_match("/[a-z][a-zA-Z]{3}/", $name)) {
-            // Ancillary chunks
-            return true;
-        }
-        return false;
-    }
-
-    public static function fromBin(DataInputStream $in)
-    {
-        if ($in -> isEof()) {
-            return null;
-        }
-        $lenData = $in -> read(4);
-        $len = unpack("N", $lenData)[1];
-        $type = $in -> read(4);
-        if (!self::isValidChunkName($type)) {
-            // In case this is not a real chunk, we don't want
-            // to use random binary data in error messages later.
-            throw new Exception("Bad chunk name");
-        }
-        $data = $in -> read($len);
-        $crcData = $in -> read(4);
-        $crc = unpack("N", $crcData)[1];
-        $chunk = new PngChunk($type, $data);
-        if ($crc != $chunk -> getCrc()) {
-            // Refuse to return chunk with bad checksum
-            throw new Exception("CRC did not match on $type chunk");
-        }
-        return $chunk;
-    }
-  
-    public function toString()
-    {
-        return $this -> type . " chunk";
-    }
-}
-
-class PngHeader
-{
-    const HEADER_SIZE = 13;
-    const COLOR_TYPE_MONOCHROME = 0;
-    const COLOR_TYPE_RGB = 2;
-    const COLOR_TYPE_INDEXED = 3;
-    const COLOR_TYPE_MONOCHROME_ALPHA = 4;
-    const COLOR_TYPE_RGBA = 6;
-
-    const COMPRESSION_DEFLATE = 0;
-
-    const INTERLACE_NONE = 0;
-    const INTERLACE_ADAM7 = 1;
-
-    private $width;
-    private $height;
-    private $bitDepth;
-    private $colorType;
-    private $compression;
-    private $filter;
-    private $interlace;
-
-    public function __construct(int $width, int $height, int $bitDepth, int $colorType, int $compression, int $filter, int $interlace)
-    {
-      // Image dimensions
-        if ($width < 1 || $width > 2147483647 ||
-        $height < 1 || $height > 2147483647) {
-            throw new Exception("Invalid image dimensions");
-        }
-        $this -> width = $width;
-        $this -> height = $height;
-      // Color type & bit depth
-      // - Only some combinations of bit depth and colorType are valid
-        $this -> bitDepth = $bitDepth;
-        $this -> colorType = $colorType;
-      // Compression
-        if ($compression != PngHeader::COMPRESSION_DEFLATE) {
-            throw new Exception("Compression type not supported");
-        }
-        $this -> compression = $compression;
-      // Filter type set
-        if ($filter != 0) {
-            throw new Exception("Filter type set not supported");
-        }
-        $this -> filter = $filter;
-      // Interlace method
-        if ($interlace != PngHeader::INTERLACE_NONE &&
-        $interlace != PngHeader::INTERLACE_ADAM7) {
-            throw new Exception("Interlace method not supported");
-        }
-        $this -> interlace = $interlace;
-    }
-
-    public static function fromChunk(PngChunk $chunk)
-    {
-        $chunkData = $chunk -> getData();
-        $chunkLen = strlen($chunkData);
-        if ($chunkLen !== PngHeader::HEADER_SIZE) {
-            throw new Exception("Header must be " . PngHeader::HEADER_SIZE . " bytes, but got $chunkLen bytes.");
-        }
-      // Unpack binary
-        $dataItems = unpack("Nwidth/Nheight/CbitDepth/CcolorType/Ccompression/Cfilter/Cinterlace", $chunkData);
-      // Construct
-        return new PngHeader($dataItems['width'], $dataItems['height'], $dataItems['bitDepth'], $dataItems['colorType'], $dataItems['compression'], $dataItems['filter'], $dataItems['interlace']);
-    }
-
-    public function toString()
-    {
-        return "Image dimensions " . $this -> width . " x " . $this -> height .
-          ", bitDepth " . $this -> bitDepth .
-          ", colorType " . $this -> colorType .
-          ", compression " . $this -> compression .
-          ", filter " . $this -> filter .
-          ", interlace " . $this -> interlace;
-    }
-
-    public function allowsPalette()
-    {
-        return $this -> requiresPalette() ||
-          $this -> colorType === PngHeader::COLOR_TYPE_RGB ||
-          $this -> colorType === PngHeader::COLOR_TYPE_RGBA;
-    }
-
-    public function requiresPalette()
-    {
-        return $this -> colorType === PngHeader::COLOR_TYPE_INDEXED;
-    }
-
-    public function getWidth()
-    {
-        return $this -> width;
-    }
-
-    public function getHeight()
-    {
-        return $this -> height;
-    }
-
-    public function getBitDepth()
-    {
-        return $this -> bitDepth;
-    }
-
-    public function getColorType()
-    {
-        return $this -> colorType;
-    }
-
-    public function getCompresssion()
-    {
-        return $this -> compresssion;
-    }
-
-    public function getFilter()
-    {
-        return $this -> filter;
-    }
-
-    public function getInterlace()
-    {
-        return $this -> interlace;
-    }
 }
 
 const PNG_SIGNATURE="\x89\x50\x4E\x47\x0D\x0A\x1A\x0A";
